Use double precision for simulation translation

+++ Engine2/Simulation Runtime/Engine/Protocol/PSystem.swift
@@ -4,7 +4,8 @@
 /// `Engine`; they are not top-level runtimes. Implementations should iterate
 /// component stores directly and use in-place component updates for existing
 /// rows. The mutating requirement also permits value-type systems to retain
-/// deliberately scoped scheduling state.
+/// deliberately scoped scheduling state. `deltaTime` is the invocation's
+/// simulation interval in double-precision seconds.
 protocol PSystem {
-    mutating func update(world: inout World, deltaTime: Float)
+    mutating func update(world: inout World, deltaTime: Double)
 }
