+++ Engine2/Simulation Runtime/Engine/System/Input/System/SInputCleanup.swift
@@ -1,6 +1,6 @@
 /// Clears one-step input deltas after every fixed input step.
 struct SInputCleanup: PSystem {
-    mutating func update(world: inout World, deltaTime: Float) {
+    mutating func update(world: inout World, deltaTime: Double) {
         world.input.clearTransientInput()
     }
 }
