+++ Engine2/Simulation Runtime/Engine/System/Position/System/SMovement.swift
@@ -1,14 +1,16 @@
 import OSLog
 
-/// Applies one frame of accumulated motion by first updating velocity and then
-/// advancing position from the new velocity.
+/// Integrates one scheduled interval of accumulated translational motion with symplectic Euler.
+///
+/// The system updates velocity before position, then clears the interval-local
+/// accumulator. Both authoritative values use double-precision SI units.
 struct SMovement: PSystem {
     private static let signposter = OSSignposter(
         subsystem: "Engine2",
         category: "SMovement"
     )
 
-    mutating func update(world: inout World, deltaTime: Float) {
+    mutating func update(world: inout World, deltaTime: Double) {
         let signpostState = Self.signposter.beginInterval("SMovement.update")
         defer {
             Self.signposter.endInterval("SMovement.update", signpostState)
@@ -22,7 +24,7 @@ struct SMovement: PSystem {
                 continue
             }
 
-            var updatedPosition: SIMD3<Float>?
+            var updatedPosition: SIMD3<Double>?
             world.motionComponents.update(for: entity) { motion in
                 // Continuous acceleration scales with `deltaTime`; impulse is an immediate
                 // velocity delta. Position then advances using the updated velocity.
